updete TC_006, TC_011
